Create wp_tag_linker.php
A simple WP plugin that highlights a maximum of five tags in the text of a post. Select from a list of tags that you need to manually specify in the settings. Also in the settings you can specify the color of the link and the possibility to underline it.